% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk.R
\name{CppFlags}
\alias{CppFlags}
\title{Compiler flags for packages linking against VTK}
\usage{
CppFlags()
}
\value{
A single character string of compiler flags, written to stdout (so
that it can be captured by shell command substitution in \code{configure}) and
returned invisibly.
}
\description{
Returns the C pre-processor flags (\code{-I} paths) required to compile C++ code
that includes VTK headers.  Intended to be called from a downstream
package's \code{configure} or \code{configure.win} script:
}
\details{
\if{html}{\out{<div class="sourceCode sh">}}\preformatted{VTK_CPPFLAGS="$("$\{R_HOME\}/bin/Rscript" --vanilla -e "rvtk::CppFlags()")"
}\if{html}{\out{</div>}}
}
\examples{
flags <- CppFlags()
}
