% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvtk-package.R
\docType{package}
\name{rvtk-package}
\alias{rvtk}
\alias{rvtk-package}
\title{rvtk: Bindings for the Visualization Toolkit ('VTK')}
\description{
Provides pre-compiled static 'VTK' libraries and headers so that downstream R packages can link against the Visualization Toolkit without requiring users to install 'VTK' manually. On all platforms the package first honours a user-supplied 'VTK_DIR' environment variable. On macOS it then tries 'Homebrew', followed by 'pkg-config'. On Linux it tries 'pkg-config' and well-known system prefixes ('/usr', '/usr/local'). If no suitable system installation is found on macOS or Linux, pre-built static libraries are downloaded automatically from the package's GitHub releases. On Windows, pre-built static libraries compiled with the 'Rtools45' 'x86_64-w64-mingw32.static.posix' toolchain are always downloaded from the same source. Note that on Windows the modules 'VTK_IONetCDF', 'VTK_IOHDF', 'VTK_GeovisCore', and 'VTK_RenderingCore' are disabled because 'netcdf' and 'libproj' are not available in the 'Rtools45' 'static.posix' sysroot. Downstream packages can declare 'Imports: rvtk' and obtain the correct compiler and linker flags at install time via rvtk::CppFlags() and rvtk::LdFlagsFile().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/astamm/rvtk}
  \item Report bugs at \url{https://github.com/astamm/rvtk/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

}
\keyword{internal}
