% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk.R
\name{CppFlags}
\alias{CppFlags}
\title{Compiler flags for packages linking against VTK}
\usage{
CppFlags()
}
\value{
A single character string of compiler flags, printed to stdout (so
that it can be captured by \verb{$(shell ...)} in a \code{Makefile}).
}
\description{
Returns the C pre-processor flags (\code{-I} paths) required to compile C++ code
that includes VTK headers.  Intended to be called from a downstream
package's \code{src/Makevars} or \code{src/Makevars.win}:
}
\details{
\if{html}{\out{<div class="sourceCode makefile">}}\preformatted{PKG_CPPFLAGS = $(shell "$(R_HOME)/bin$(R_ARCH_BIN)/Rscript" -e "rvtk::CppFlags()")
}\if{html}{\out{</div>}}
}
