% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk.R
\name{LdFlags}
\alias{LdFlags}
\title{Linker flags for packages linking against VTK}
\usage{
LdFlags()
}
\value{
A single character string of linker flags, printed to stdout (so
that it can be captured by \verb{$(shell ...)} in a \code{Makefile}).
}
\description{
Returns the linker flags (\code{-L} paths and \code{-l} library names) required to
link C++ code against VTK.  Intended to be called from a downstream
package's \code{src/Makevars} or \code{src/Makevars.win}:
}
\details{
\if{html}{\out{<div class="sourceCode makefile">}}\preformatted{PKG_LIBS = $(shell "$(R_HOME)/bin$(R_ARCH_BIN)/Rscript" -e "rvtk::LdFlags()")
}\if{html}{\out{</div>}}
}
